+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Not yet released
 * Add `WITHIN GROUP` syntax support for ordered set-aggregate functions
 * Add SQL standard compliant `LISTAGG` ordered set-aggregate function as alternative to `GROUP_CONCAT`
 * Add `PERCENTILE_CONT`, `PERCENTILE_DISC` and `MODE` ordered set-aggregate functions
+* Add JPA Criteria support for aggregate `FILTER` clause, window functions and ordered set aggregate functions
 
 ### Bug fixes